Move offline ddoc logic to the bootstrapper js code.
