--- webapp/src/ts/services/offline-ddocs/design-doc.ts
@@ -1,11 +1,0 @@
-export interface DesignDoc {
-  readonly _id: `_design/${string}`;
-  readonly _rev?: string;
-  readonly views: {
-    [key: string]: {
-      map: string;
-    };
-  };
-}
-
-export const packageView = ({ map }: { map: Function }) => ({ map: map.toString() });
