+++ webapp/tests/karma/ts/app.component.spec.ts
@@ -151,7 +151,6 @@ describe('AppComponent', () => {
       isOnlineOnly: sinon.stub()
     };
     dbSyncService = {
-      init: sinon.stub().resolves(),
       addUpdateListener: sinon.stub(),
       isEnabled: sinon.stub().returns(false),
       sync: sinon.stub(),
@@ -500,7 +499,6 @@ describe('AppComponent', () => {
       }]);
       expect(globalActions.updateReplicationStatus.getCall(1).args).to.deep.equal([{disabled: true}]);
       expect(dbSyncService.subscribe.callCount).to.equal(1);
-      expect(dbSyncService.init.calledOnceWithExactly()).to.be.true;
     });
 
     it('should sync db if enabled', async () => {
@@ -521,7 +519,6 @@ describe('AppComponent', () => {
 
       expect(dbSyncService.sync.callCount).to.equal(1);
       expect(dbSyncService.subscribe.callCount).to.equal(1);
-      expect(dbSyncService.init.calledOnceWithExactly()).to.be.true;
     });
 
     it('should set dbSync replication status in subcription callback', async () => {
