--- webapp/tests/mocha/ts/.mocharc.js
@@ -1,8 +1,0 @@
-const chaiAsPromised = require('chai-as-promised');
-const chai = require('chai');
-chai.use(chaiAsPromised);
-
-module.exports = {
-  spec: 'tests/mocha/ts/**/*.spec.ts',
-  require: 'ts-node/register',
-};
