+++ webapp/tsconfig.spec.json
@@ -16,7 +16,6 @@
   ],
   "include": [
     "tests/karma/**/*.spec.ts",
-    "tests/karma/**/*.d.ts",
-    "tests/mocha/ts/**/*.spec.ts",
+    "tests/karma/**/*.d.ts"
   ]
 }
